fix: Resolve App initialization errors in examples and fix warning

+++ src/database/MongoClient.cpp
@@ -101,7 +101,7 @@ void MongoClient::InsertAwaiter::await_suspend(std::coroutine_handle<> h) {
         mongoc_collection_t* collection = mongoc_client_get_collection(mclient, client.get_config().dbname.c_str(), client.get_config().collection_name.c_str());
         
         bson_error_t error;
-        bson_t* doc = bson_new_from_json(reinterpret_cast<const uint8_t*>(doc_json.c_str()), doc_json.length(), &error);
+        bson_t* doc = bson_new_from_json(reinterpret_cast<const uint8_t*>(doc_json.c_str()), static_cast<ssize_t>(doc_json.length()), &error);
         
         if (!doc) {
             error_msg = std::string("BSON Parse Error: ") + error.message;
